backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
bitarray==2.9.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
constantly==23.10.4
cryptography==43.0.1
cytoolz==1.0.0
daphne==4.1.2
Django==4.2.16
django-celery-results==2.5.1
django-cors==0.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
ecdsa==0.19.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
factory_boy==3.3.1
Faker==30.3.0
frozenlist==1.4.1
git-filter-repo==2.47.0
gunicorn==23.0.0
hexbytes==1.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
packaging==24.1
parsimonious==0.10.0
prompt_toolkit==3.0.50
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.3.0
pyunormalize==16.0.0
redis==5.2.1
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.1
stripe==11.1.0
toolz==1.0.0
Twisted==24.11.0
txaio==23.1.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
web3==7.7.0
websockets==13.1
whitenoise==6.7.0
yarl==1.14.0
zope.interface==7.2
